1.完成主界面分类详情内容的请求

+++ app/src/main/java/com/guoj/store/model/bean/HomePagerContent.kt
@@ -1,7 +1,7 @@
 package com.guoj.store.model.bean
 
 data class HomePagerContent(
-    val code: Int,
+    val code: Long,
     val `data`: List<Data>,
     val message: String,
     val success: Boolean
@@ -13,31 +13,31 @@ data class HomePagerContent(
 }
 
 data class Data(
-    val category_id: Int,
+    val category_id: Long,
     val category_name: Any,
     val click_url: String,
     val commission_rate: String,
-    val coupon_amount: Int,
+    val coupon_amount: Long,
     val coupon_click_url: String,
     val coupon_end_time: String,
     val coupon_info: Any,
-    val coupon_remain_count: Int,
+    val coupon_remain_count: Long,
     val coupon_share_url: String,
     val coupon_start_fee: String,
     val coupon_start_time: String,
-    val coupon_total_count: Int,
+    val coupon_total_count: Long,
     val item_description: String,
     val item_id: Long,
-    val level_one_category_id: Int,
+    val level_one_category_id: Long,
     val level_one_category_name: String,
     val nick: String,
     val pict_url: String,
-    val seller_id: Int,
+    val seller_id: Long,
     val shop_title: String,
     val small_images: SmallImages,
     val title: String,
-    val user_type: Int,
-    val volume: Int,
+    val user_type: Long,
+    val volume: Long,
     val zk_final_price: String
 
 ) {
